Add support for System.Drawing on Windows.
This change adds a new library, System.Drawing.Common, which supports a
large subset of the System.Drawing API. It is a direct port of the code
from the .NET Framework codebase. The code has been cleaned, formatted,
and sanitized to match the code style of corefx, but is otherwise very
similar to the .NET Framework version. There are a few key differences:

* TypeConverter's have been omitted for now. If we want to add these to
  .NET Core, we should add them to the TypeConverter library, rather than
  System.Drawing.Common itself.
* Some parts of the library do not respond to "System Events", because
  they are not yet implemented in .NET Core. This means that fonts and
  colors will not automatically update when the system defaults change,
  as they do in .NET Framework. The code is still there, but behind the
  feature flag "FEATURE_SYSTEM_EVENTS".
* Various attributes have been removed from the codebase, because they are
  no longer applicable. Examples:
  * Designer-related attributes (DefaultProperty, DefaultEvent, Editor,
    etc.)
  * CAS-related attributes. CAS-related method calls (Security demands and
    asserts) have also been removed.

There is still further cleanup that can be done in order to match the
style and structure of corefx. For example, the Interop code should be
reorganized to match other BCL libraries.

NOTE: This implementation will only work on Windows, and not in UWP. This
is a strictly compatibility-driven feature. Any changes made to it should
not significantly diverge from the .NET Framework implementation.
Libraries should only take a dependency on System.Drawing.Common if they
are relying on legacy components from .NET Framework; newly-written code
and features should rely on modern alternatives.