Fix issue in shared drawing component
In cases were we return the default bitmap
in the ToolboxBitmapAttribute class we were handing
out a shared image file which was getting disposed
in some cases which causes issue if others access it.

To fix this we are not cloning the image before returning
it so the shared image doesn't get destroyed for others.