Clean style in Graphics/GraphicsPath
Make a consistency pass through Graphics & GraphicsPath.

Plan is to follow up with a change that removes the rest of the unnecessary allocations.

This change focuses on the Windows specific code. Some Unix code will be able to be unified with the Windows code once further changes happen. Rather than reformat the Unix code preemptively, leaving for final follow-up.